// src/app/models/company-update-request.ts
export interface CompanyUpdateRequest {
    oldPassword: string;
    cep: string;
    cnpj: string;
    name: string;
    email: string;
    phone: string;
    password: string;
    street: string;
    neighborhood: string;
    city: string;
    state: string;
    complement: string;
    number: string;
  }
  